remove image if product is deleted

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
-
+const fileHelper = require('./../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -131,6 +131,7 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             product.description = updatedDesc;
@@ -149,7 +150,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
